requirements.txt
streamlit
openai-whisper
transformers==4.41.2
torch
yt-dlp
sentencepiece
accelerate
sentence-transformers
faiss-cpu
gTTS
python-dotenv
openai
fastapi
uvicorn[standard]
python-multipart
pytest
pytest-cov
pytest-mock
httpx
flake8